requirements.txt
Flask
Pillow
python-telegram-bot>=20.0
